refactor: supabase config environment
