--- libs/environments/src/supabase/config.example.json
@@ -1,3 +1,0 @@
-{
-  "anonKey": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.my_supabase_anon_key"
-}
